maint #341 extract exception classes from misc.py into exceptions.py
Create hklpy2/exceptions.py with all 10 custom exception classes:
Hklpy2Error, ConfigurationError, ConstraintsError, CoreError,
DiffractometerError, LatticeError, NoForwardSolutions, ReflectionError,
SampleError, SolverError.

Remove the class definitions from misc.py (no re-export shim).
Update all internal imports across source, backend, block, and test
files to import directly from hklpy2.exceptions. misc.py itself
imports NoForwardSolutions and SolverError from exceptions for its
own use in pick_*_solution() and get_solver().

Agent: OpenCode (claudesonnet46)
