+++ src/hklpy2/__init__.py
@@ -65,7 +65,7 @@ from .diffract import diffractometer_class_factory  # noqa: E402, F401, F403
 from .incident import A_KEV  # noqa: E402, F401
 from .misc import SOLVER_ENTRYPOINT_GROUP  # noqa: E402, F401
 from .misc import ConfigurationRunWrapper  # noqa: E402, F401
-from .misc import SolverError  # noqa: E402, F401
+from .exceptions import SolverError  # noqa: E402, F401
 from .misc import creator_from_config  # noqa: E402, F401
 from .misc import get_run_orientation  # noqa: E402, F401
 from .misc import get_solver  # noqa: E402, F401
