+++ src/hklpy2/blocks/reflection.py
@@ -19,8 +19,8 @@ from typing import Optional
 from typing import Union
 
 from ..misc import INTERNAL_LENGTH_UNITS
-from ..misc import ConfigurationError
-from ..misc import ReflectionError
+from ..exceptions import ConfigurationError
+from ..exceptions import ReflectionError
 from ..misc import check_value_in_list
 from ..misc import compare_float_dicts
 from ..misc import convert_units
