+++ src/hklpy2/diffract.py
@@ -46,7 +46,7 @@ from .incident import WavelengthXray
 from .misc import DEFAULT_DIGITS
 from .misc import INTERNAL_ANGLE_UNITS
 from .misc import MISSING_HEADER_KEY_MSG
-from .misc import DiffractometerError
+from .exceptions import DiffractometerError
 from .misc import load_yaml_file
 from .misc import pick_first_solution
 from .misc import roundoff
