+++ src/hklpy2/tests/test_i190.py
@@ -16,7 +16,7 @@ from numpy.testing import assert_almost_equal
 
 from ..blocks.lattice import SI_LATTICE_PARAMETER
 from ..diffract import creator
-from ..misc import NoForwardSolutions
+from ..exceptions import NoForwardSolutions
 
 
 @pytest.fixture()
